requirements.txt
cryptography==44.0.2
datetime==5.5
numpy==1.26.4
pandas==2.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scipy==1.12.0
ta==0.11.0  # Technical analysis library
urllib3==2.3.0
websockets==14.1
